Rename agent alive API types and drop chatConnection indirection.
Use AgentAliveResponse/fetchAgentAlive and agentAlivePoll.ts naming.
Move WebSocket close reset into handleChatConnectionLost on the poll module
instead of register/notify handlers in chatConnection.ts.
