+++ frontend/src/services/websocket/chat.ts
@@ -1,7 +1,7 @@
 import type { WSMessage } from '../../types';
 import type { ChatSocket } from './types';
 import { readDataSourceAuthorization } from '../dataSourceAuth';
-import { notifyChatConnectionLost } from '../../stores/chatConnection';
+import { handleChatConnectionLost } from '../../stores/agentAlivePoll';
 import { createReconnectingSocket, getWsBase } from './reconnecting';
 
 const chatSockets = new Map<string, ChatSocket>();
@@ -40,7 +40,7 @@ export function getChatSocket(projectId: string): ChatSocket {
         handlers.forEach((h) => h(msg));
       } catch {}
     },
-    notifyChatConnectionLost,
+    handleChatConnectionLost,
     () => {
       if (typeof window !== 'undefined') {
         import('../../stores/errors').then(({ useErrorStore }) => {
