--- frontend/src/stores/chatConnection.ts
@@ -1,9 +1,0 @@
-let connectionLostHandler: (() => void) | null = null;
-
-export function registerChatConnectionLostHandler(handler: () => void): void {
-  connectionLostHandler = handler;
-}
-
-export function notifyChatConnectionLost(): void {
-  connectionLostHandler?.();
-}
